# apps/backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Arenda Planform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /admin/tariffs:
    get:
      operationId: listAdminTariffs
      security:
        - sessionCookie: []
      responses:
        '200':
          description: All tariffs, including hidden ones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTariffsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createAdminTariff
      description: |
        Creates a tariff plan (issue #256). The name comes from the closed
        TariffName vocabulary the frozen user contract pins; a name that
        already exists answers 409.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateTariffRequest'
      responses:
        '201':
          description: The created tariff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTariff'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: A tariff with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/tariffs/{tariffId}:
    put:
      operationId: updateAdminTariff
      description: |
        Updates a tariff's prices, active-property limit and activity
        (issue #256). Hiding (isActive=false) removes the plan from the
        user-facing listing and blocks new selection; existing subscriptions,
        renewals and the payment history keep referencing it. Paid periods
        already granted are not repriced — the next renewal charges the
        updated price.
      security:
        - sessionCookie: []
      parameters:
        - name: tariffId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateTariffRequest'
      responses:
        '200':
          description: The updated tariff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTariff'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/subscription/payments:
    get:
      operationId: listAdminSubscriptionPayments
      security:
        - sessionCookie: []
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SubscriptionPaymentStatus'
        - name: user_phone
          in: query
          required: false
          description: Exact match on the user's phone number.
          schema:
            type: string
        - name: subscription_status
          in: query
          required: false
          description: Filters payments by the payer's current subscription status (issue #254).
          schema:
            $ref: '#/components/schemas/SubscriptionStatus'
        - name: sort
          in: query
          required: false
          description: 'Sort field (camelCase). Allowed: createdAt, amountKopecks, status. Defaults to createdAt descending.'
          schema:
            type: string
        - name: order
          in: query
          required: false
          description: 'Sort direction: asc or desc (default desc).'
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Admin subscription payments list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionPaymentsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/subscription/payments/{paymentId}:
    get:
      operationId: getAdminSubscriptionPayment
      security:
        - sessionCookie: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin subscription payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSubscriptionPayment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/subscription/payments/{paymentId}/refund:
    post:
      operationId: refundSubscriptionPayment
      security:
        - sessionCookie: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment refunded
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/subscription/payments/{paymentId}/sync:
    post:
      operationId: syncSubscriptionPayment
      security:
        - sessionCookie: []
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment synchronized
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users:
    get:
      operationId: listAdminUsers
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: phone
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [owner, admin]
        - name: subscription_status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SubscriptionStatus'
        - name: sort
          in: query
          required: false
          description: 'Sort field (camelCase). Allowed: createdAt, updatedAt. Defaults to createdAt descending.'
          schema:
            type: string
        - name: order
          in: query
          required: false
          description: 'Sort direction: asc or desc (default desc).'
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Admin users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{id}/properties:
    get:
      operationId: listAdminUserProperties
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived, all]
            default: all
        - name: sort
          in: query
          required: false
          description: 'Sort field (camelCase). Allowed: name, createdAt, updatedAt, status. Defaults to updatedAt descending.'
          schema:
            type: string
        - name: order
          in: query
          required: false
          description: 'Sort direction: asc or desc (default desc).'
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Admin user properties list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPropertiesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{id}/leases:
    get:
      operationId: listAdminUserLeases
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          required: false
          description: 'Sort field (camelCase). Allowed: startDate, updatedAt, status, rentAmountKopecks. Defaults to updatedAt descending.'
          schema:
            type: string
        - name: order
          in: query
          required: false
          description: 'Sort direction: asc or desc (default desc).'
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Admin user leases list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLeasesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/users/{id}/tenant-contacts:
    get:
      operationId: listAdminUserTenantContacts
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          required: false
          description: 'Sort field (camelCase). Allowed: name, updatedAt. Defaults to updatedAt descending.'
          schema:
            type: string
        - name: order
          in: query
          required: false
          description: 'Sort direction: asc or desc (default desc).'
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Admin user tenant contacts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminTenantContactsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/properties/{id}:
    get:
      operationId: getAdminProperty
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin property details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/leases/{id}:
    get:
      operationId: getAdminLease
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Admin lease details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLeaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /client-errors:
    post:
      operationId: reportClientError
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientErrorReport'
      responses:
        '204':
          description: Error report accepted
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: >
        The actual cookie name depends on the `cookieSecure` setting:
        `session_id` when insecure and `__Host-session_id` when secure.
...
  schemas:
...
    AdminTariffsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminTariff'
...
    SubscriptionStatus:
      type: string
      enum: [active, grace, cancelled]
    SubscriptionPaymentStatus:
      type: string
      enum: [pending, succeeded, failed, refunded, partial_refunded, refunding]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'